+++ src/untangled/client/impl/network.cljs
@@ -7,74 +7,85 @@
     [cljs.core.async.macros :refer [go]])
   (:import [goog.net XhrIo EventType]))
 
+(declare make-untangled-network)
+
+(defn make-xhrio []
+  (XhrIo.))
+
 (defprotocol UntangledNetwork
   (send [this edn ok-callback error-callback]
-    "Send method, transmits EDN to the server and gets an EDN response. Calls result-callback with that response,
-    or a map with key `:error` on errors. optional options may include `:headers`, but you may NOT override content
-    type."))
+        "Send method, transmits EDN to the server and gets an EDN response. Calls result-callback with that response,
+        or a map with key `:error` on errors. optional options may include `:headers`, but you may NOT override content
+        type. The method CANNOT be used for parallel network requests.")
+  (start [this complete-app]
+    "Starts the network, passing in the app for any components that may need it."))
 
 (defprotocol IXhrIOCallbacks
-  (response-ok [this] "Called by XhrIo on OK")
-  (response-error [this] "Called by XhrIo on ERROR"))
+  (response-ok [this xhrio ok-cb] "Called by XhrIo on OK")
+  (response-error [this xhrio err-cb] "Called by XhrIo on ERROR"))
 
 (defn parse-response [xhr-io]
   (ct/read (t/reader {:handlers {"f" (fn [v] (js/parseFloat v))}}) (.getResponseText xhr-io)))
 
-(defrecord Network [xhr-io url error-callback valid-data-callback request-transform global-error-callback]
+(defrecord Network [url request-transform global-error-callback complete-app]
   IXhrIOCallbacks
-  (response-ok [this]
+  (response-ok [this xhr-io valid-data-callback]
     ;; Implies:  everything went well and we have a good response
     ;; (i.e., got a 200).
-    (let [query-response (parse-response xhr-io)]
-      (when (and query-response @valid-data-callback) (@valid-data-callback query-response))))
+    (try
+      (let [query-response (parse-response xhr-io)]
+        (when (and query-response valid-data-callback) (valid-data-callback query-response)))
+      (finally (.dispose xhr-io))))
 
-  (response-error [this]
+  (response-error [this xhr-io error-callback]
     ;; Implies:  request was sent.
     ;; *Always* called if completed (even in the face of network errors).
     ;; Used to detect errors.
-    (letfn [(log-and-dispatch-error [str error]
-              ;; note that impl.application/initialize will partially apply the
-              ;; app-state as the first arg to global-error-callback
-              (log/error str)
-              (@global-error-callback error)
-              (@error-callback error))]
-
-      (if (zero? (.getStatus xhr-io))
-        (log-and-dispatch-error
-          (str "UNTANGLED NETWORK ERROR: No connection established.")
-          {:type :network})
-        (log-and-dispatch-error
-          (str "SERVER ERROR CODE: " (.getStatus xhr-io))
-          (parse-response xhr-io)))))
+    (try
+      (let [status (.getStatus xhr-io)
+            log-and-dispatch-error (fn [str error]
+                                     ;; note that impl.application/initialize will partially apply the
+                                     ;; app-state as the first arg to global-error-callback
+                                     (log/error str)
+                                     (when @global-error-callback
+                                         (@global-error-callback status error))
+                                     (error-callback error))]
+        (if (zero? status)
+          (log-and-dispatch-error
+            (str "UNTANGLED NETWORK ERROR: No connection established.")
+            {:type :network})
+          (log-and-dispatch-error
+            (str "SERVER ERROR CODE: " status)
+            (parse-response xhr-io))))
+      (finally (.dispose xhr-io))))
 
   UntangledNetwork
   (send [this edn ok err]
-    (let [headers {"Content-Type" "application/transit+json"}
+    (let [xhrio (make-xhrio)
+          headers {"Content-Type" "application/transit+json"}
           {:keys [request headers]} (cond
                                       request-transform (request-transform {:request edn :headers headers})
                                       :else {:request edn :headers headers})
           post-data (ct/write (t/writer) request)
           headers (clj->js headers)]
-      (reset! error-callback (fn [e] (err e)))
-      (reset! valid-data-callback (fn [resp] (ok resp)))
-      (.send xhr-io url "POST" post-data headers))))
+      (.send xhrio url "POST" post-data headers)
+      (events/listen xhrio (.-SUCCESS EventType) #(response-ok this xhrio ok))
+      (events/listen xhrio (.-ERROR EventType) #(response-error this xhrio err))))
+
+  (start [this app]
+    (assoc this :complete-app app)))
 
 (defn make-untangled-network [url & {:keys [request-transform global-error-callback]}]
-  (let [xhrio (XhrIo.)
-        rv (map->Network {:xhr-io                xhrio
-                          :url                   url
-                          :request-transform     request-transform
-                          :global-error-callback (atom global-error-callback)
-                          :valid-data-callback   (atom nil)
-                          :error-callback        (atom nil)})]
-    (events/listen xhrio (.-SUCCESS EventType) #(response-ok rv))
-    (events/listen xhrio (.-ERROR EventType) #(response-error rv))
-    rv))
+  (map->Network {:url                   url
+                 :request-transform     request-transform
+                 :global-error-callback (atom global-error-callback)
+                 }))
 
-(defrecord MockNetwork []
+(defrecord MockNetwork [complete-app]
   UntangledNetwork
   (send [this edn ok err]
-    (log/info "Ignored (mock) Network request " edn)))
+    (log/info "Ignored (mock) Network request " edn))
+  (start [this app]
+    (assoc this :complete-app app)))
 
-(defn mock-network [] (MockNetwork.))
-
+(defn mock-network [] (map->MockNetwork {}))