+++ CHANGELOG.md
@@ -25,9 +25,24 @@
 - Renamed everything in the internals that was prefixed app/ to untangled/
 - Added global server error handler.
 - Added fallback support for load-data, load-field, etc.
-- Refactored networking send for better clarity 
+- Refactored networking send for better clarity
 - Fixed bug on mark/sweep of missing query results. It was being applied to mutations instead of reads. Added tests for this that need a bit more work.
 
 0.4.8
 -----
 - Fixed tempid rewrite regression
+
+0.4.9
+-----
+- Removed old logging code (use untangled.client.logging instead)
+- Added support for parallel lazy loading
+- Added `(start [this app])` to the `UntangledNetwork` protocol.
+- Log-app-state now requires the atom containing a mounted untangled client, define it in the user namespace like so:
+```
+(defonce app (atom (uc/new-untangled-client ... )))
+(swap! app uc/mount RootComponent "app-div")
+(def log-app-state (partial util/log-app-state app))
+```
+- `global-error-callback` now expectes an arity 2 function. First param is the status and the second is the response.
+- Fixed bug that closed over tempids in network callbacks
+- Fixed bug in path-optimized union query parsing